+++ common-libs/did-auth-lib/src/DidAuthService/DidAuthServerService.ts
@@ -1,19 +1,22 @@
 import { Affinidi } from '@affinidi/common'
 import { JwtService } from '@affinidi/tools-common'
-import { parse } from 'did-resolver'
 import { DidAuthResponseToken } from './DidAuthResponseToken'
 import { DEFAULT_REQUEST_TOKEN_VALID_IN_MS } from '../shared/constants'
 import Signer from '../shared/Signer'
 
 export default class DidAuthServerService {
-  constructor(private readonly _did: string, private readonly _signer: Signer, private readonly _affinidi: Affinidi) {}
+  constructor(
+    private readonly _verifierDid: string,
+    private readonly _signer: Signer,
+    private readonly _affinidi: Affinidi,
+  ) {}
 
   async createDidAuthRequestToken(audienceDid: string, expiresAt?: number): Promise<string> {
     const jwtType = 'DidAuthRequest'
     const NOW = Date.now()
 
     const jwtObject: any = await JwtService.buildJWTInteractionToken(null, jwtType, null)
-    jwtObject.payload.aud = parse(audienceDid).did
+    jwtObject.payload.aud = audienceDid
     jwtObject.payload.exp = expiresAt > NOW ? expiresAt : NOW + DEFAULT_REQUEST_TOKEN_VALID_IN_MS
     jwtObject.payload.createdAt = NOW
 
@@ -29,9 +32,7 @@ export default class DidAuthServerService {
     await this._affinidi.validateJWT(didAuthRequestToken.toString())
     await this._affinidi.validateJWT(didAuthResponseToken.toString(), didAuthRequestToken.toString())
 
-    const verifierDid = parse(this._did).did
-
-    if (didAuthRequestToken.iss !== verifierDid) {
+    if (didAuthRequestToken.iss !== this._verifierDid) {
       throw new Error('Issuer of request is not valid')
     }