+++ common-libs/did-auth-lib/CHANGELOG.md
@@ -4,7 +4,8 @@
 * `node-fetch` dependency replaced with `internal-api-clients` dependency.
 * In order to simplify dependencies:
   * `DidAuthClientService` constructor now accepts `Signer` as a parameter;
-  * `DidAuthServerService`: did, signer and `Affinidi` instance;
+  * `DidAuthServerService`: short-form verifier did, signer and `Affinidi` instance;
+    * `createDidAuthRequestToken` accepts short-form audience did.
   * `DidAuthCloudService`: `CloudWalletApiService`.
 # release 2.0.0-beta.0 (2021-10-13)
 * Code imported into monorepo from private repository